+++ src/libsac2c/constants/cv2str.c
@@ -35,9 +35,9 @@
  ******************************************************************************/
 
 #define COcv2StrTEMPLATE(type, ext, form)                                                \
-    char *COcv2Str##ext (void *src, int off, int len, int max_char)                      \
+    char *COcv2Str##ext (void *src, size_t off, size_t len, int max_char)                \
     {                                                                                    \
-        int i;                                                                           \
+        size_t i;                                                                        \
         char format[10];                                                                 \
         char *buffer;                                                                    \
         char *buffer_act;                                                                \
@@ -45,7 +45,7 @@
                                                                                          \
         DBUG_ENTER ();                                                                   \
         sprintf (format, ",%s", form);                                                   \
-        buffer = (char *)MEMmalloc ((100 + max) * sizeof (char));                   \
+        buffer = (char *)MEMmalloc ((100 + max) * sizeof (char));                        \
         buffer_act = buffer;                                                             \
                                                                                          \
         if (len > 0) {                                                                   \
@@ -85,16 +85,16 @@ COcv2StrTEMPLATE (unsigned char, UByte, "%c")
 
   /* SIMD vectors are kind of special case.
      FIXME Do the implementation.  */
-  char *COcv2StrFloatvec (void *src, int off, int len, int max_char)
+  char *COcv2StrFloatvec (void *src, size_t off, size_t len, int max_char)
 {
     char *buf = (char *)MEMmalloc (1024);
     char *p = buf;
     char *end = p + 1024;
     size_t n_left = (size_t) (end - p);
 
-    p += snprintf (p,  n_left, "floatvec<%d>[", len);
+    p += snprintf (p,  n_left, "floatvec<%zu>[", len);
 
-    for (int i = 0; i < len; i++){
+    for (size_t i = 0; i < len; i++){
         n_left = (size_t) (end - p);
         if (i < 3)
             p += snprintf (p, n_left, "[%f,...]", *(float *)&((floatvec *)src)[i + off]);
@@ -117,7 +117,7 @@ COcv2StrTEMPLATE (unsigned char, UByte, "%c")
  */
 
 char *
-COcv2StrDummy (void *src, int off, int len, int max_char)
+COcv2StrDummy (void *src, size_t off, size_t len, int max_char)
 {
     DBUG_ENTER ();
     DBUG_UNREACHABLE ("COcv2StrDummy called!");
