+++ src/libsac2c/constants/zipcv.c
@@ -22,8 +22,8 @@
 /******************************************************************************
  *
  * function:
- *    void COzipCvXXXYYY( void *arg1, int pos1, void *arg2, int pos2,
- *                        void *res, int res_pos)
+ *    void COzipCvXXXYYY( void *arg1, size_t pos1, void *arg2, size_t pos2,
+ *                        void *res, size_t res_pos)
  *
  * description:
  *    Functions for applying function YYY to two elements of cv's arg1 and arg2
@@ -33,8 +33,8 @@
  *
  *
  * function:
- *    void COzipCvDummyYYY( void *arg1, int pos1, void *arg2, int pos2,
- *                        void *res, int res_pos)
+ *    void COzipCvDummyYYY( void *arg1, size_t pos1, void *arg2, size_t pos2,
+ *                        void *res, size_t res_pos)
  *
  * description:
  *    Finally, we provide a dummy function for each operation YYY which should
@@ -46,8 +46,8 @@
 #define str(s) #s
 
 #define COzipCvTEMPLATE(fun, fun_ext, arg_t, arg_ext, res_t)                             \
-    void COzipCv##arg_ext##fun_ext (void *arg1, int pos1, void *arg2, int pos2,          \
-                                    void *res, int res_pos)                              \
+    void COzipCv##arg_ext##fun_ext (void *arg1, size_t pos1, void *arg2, size_t pos2,    \
+                                    void *res, size_t res_pos)                           \
     {                                                                                    \
         DBUG_ENTER ();                                                                   \
         ((res_t *)res)[res_pos] = (res_t) (((arg_t *)arg1)[pos1] fun                     \
@@ -56,8 +56,8 @@
     }
 
 #define COzipCvDUMMYTEMP(arg_ext, fun_ext)                                               \
-    void COzipCv##arg_ext##fun_ext (void *arg1, int pos1, void *arg2, int pos2,          \
-                                    void *res, int res_pos)                              \
+    void COzipCv##arg_ext##fun_ext (void *arg1, size_t pos1, void *arg2, size_t pos2,    \
+                                    void *res, size_t res_pos)                           \
     {                                                                                    \
         DBUG_ENTER ();                                                                   \
         DBUG_UNREACHABLE (str (COzipCv##arg_ext##fun_ext called !));                     \
@@ -65,8 +65,8 @@
     }
 
 #define COzipCvMINMAXTEMPLATE(fun, fun_ext, arg_t, arg_ext, res_t)                       \
-    void COzipCv##arg_ext##fun_ext (void *arg1, int pos1, void *arg2, int pos2,          \
-                                    void *res, int res_pos)                              \
+    void COzipCv##arg_ext##fun_ext (void *arg1, size_t pos1, void *arg2, size_t pos2,    \
+                                    void *res, size_t res_pos)                           \
     {                                                                                    \
         DBUG_ENTER ();                                                                   \
         ((res_t *)res)[res_pos] = (((arg_t *)arg1)[pos1])fun (((arg_t *)arg2)[pos2])     \
@@ -76,8 +76,8 @@
     }
 
 #define COzipCvUNARYTEMPLATE(fun, fun_ext, arg_t, arg_ext, res_t)                        \
-    void COzipCv##arg_ext##fun_ext (void *arg1, int pos1, void *arg2, int pos2,          \
-                                    void *res, int res_pos)                              \
+    void COzipCv##arg_ext##fun_ext (void *arg1, size_t pos1, void *arg2, size_t pos2,    \
+                                    void *res, size_t res_pos)                           \
     {                                                                                    \
         DBUG_ENTER ();                                                                   \
         ((res_t *)res)[res_pos] = (res_t) fun (((arg_t *)arg1)[pos1]);                   \
@@ -89,8 +89,8 @@
 /* Ensure that the code below and the code in runtime/essentials/prf.h match! */
 
 #define COzipCvAPLMODTEMPLATE(fun, fun_ext, arg_t, arg_ext, res_t)                       \
-    void COzipCv##arg_ext##fun_ext (void *arg1, int pos1, void *arg2, int pos2,          \
-                                    void *res, int res_pos)                              \
+    void COzipCv##arg_ext##fun_ext (void *arg1, size_t pos1, void *arg2, size_t pos2,    \
+                                    void *res, size_t res_pos)                           \
     {                                                                                    \
         arg_t x;                                                                         \
         arg_t y;                                                                         \
@@ -107,8 +107,8 @@
     }
 
 #define COzipCvABSTEMPLATE(fun, fun_ext, arg_t, arg_ext, res_t)                          \
-    void COzipCv##arg_ext##fun_ext (void *arg1, int pos1, void *arg2, int pos2,          \
-                                    void *res, int res_pos)                              \
+    void COzipCv##arg_ext##fun_ext (void *arg1, size_t pos1, void *arg2, size_t pos2,    \
+                                    void *res, size_t res_pos)                           \
     {                                                                                    \
         DBUG_ENTER ();                                                                   \
         ((res_t *)res)[res_pos] = (res_t)(((arg_t *)arg1)[pos1] < 0                      \
@@ -118,8 +118,8 @@
     }
 
 #define COzipCvTOBOOLTEMPLATE(fun, fun_ext, arg_t, arg_ext, res_t)                       \
-    void COzipCv##arg_ext##fun_ext (void *arg1, int pos1, void *arg2, int pos2,          \
-                                    void *res, int res_pos)                              \
+    void COzipCv##arg_ext##fun_ext (void *arg1, size_t pos1, void *arg2, size_t pos2,    \
+                                    void *res, size_t res_pos)                           \
     {                                                                                    \
         DBUG_ENTER ();                                                                   \
         if (0 == (((arg_t *)arg1)[pos1])) {                                              \
@@ -142,8 +142,8 @@
  * The following macro is not yet used but might prove useful in the future:
  */
 #define COzipCvTOCTEMPLATE(fun, fun_ext, arg_t, arg_ext, res_t)                          \
-    void COzipCv##arg_ext##fun_ext (void *arg1, int pos1, void *arg2, int pos2,          \
-                                    void *res, int res_pos)                              \
+    void COzipCv##arg_ext##fun_ext (void *arg1, size_t pos1, void *arg2, size_t pos2,    \
+                                    void *res, size_t res_pos)                           \
     {                                                                                    \
         DBUG_ENTER ();                                                                   \
                                                                                          \
